pyproject.toml
[tool.poetry]
name = "zadanko"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
ipython = "^8.12.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
pytest-django = "^4.5.2"
factory-boy = "^3.2.1"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "zadanko.settings"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
